Refactor Backend
