+++ server/Database/DAO.js
@@ -1,0 +1,9 @@
+const sqlite = require('sqlite3');
+
+// Open the database connection
+const db = new sqlite.Database('./Database/DB.sqlite', (err) => {
+    if (err) console.error(err.message);
+    console.log('DAO ready.');
+});
+
+export default db;
