--- server/Models/Test.js
@@ -1,14 +1,0 @@
-module.exports = {
-    getTest: () => {
-        return new Promise((resolve, reject) => {
-            console.log("GET /api/test");
-            resolve('GET /api/test')
-        });
-    },
-    postTest: (test) => {
-        return new Promise((resolve, reject) => {
-            console.log(test);
-            resolve('POST /api/test');
-        });
-    }
-};
