+++ server/Router/RouterAPI.js
@@ -1,26 +1,10 @@
 const express = require('express');
 const apiRouter = express.Router();
 
-apiRouter.use('/test', require('./RouterTest'));
-
-apiRouter.use('/proposals', require('./RouterSearch'));
-
-apiRouter.use('/application', require('./RouterApplication'));
-
-apiRouter.use('/auth', require('./RouterAuth'));
-
-apiRouter.use('/thesis', require('./RouterThesis'));
-
-apiRouter.use('/thesis', require('./ProposalRouter'));
-
-apiRouter.use('/degree', require('./RouterDegree'));
-
-apiRouter.use('/teacher', require('./RouterTeacher'));
-
-apiRouter.use('/cosupervisor', require('./RouterCoSupervisor'));
-
-apiRouter.use('/groups', require('./RouterResearchGroup'));
-
+apiRouter.use('/proposals', require('./ProposalsRouter'));
+apiRouter.use('/applications', require('./ApplicationsRouter'));
+apiRouter.use('/auth', require('./AuthRouter'));
+apiRouter.use('/utilities', require('./UtilitiesRouter'));
 
 // All the routes will be protected by the checkAuthentication middleware
 const checkAuthentication = (req, res, next) => {
