--- server/Router/RouterDegree.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const degreeRouter = express.Router();
-
-const degreeController = require('../Controllers/DegreeController');
-
-
-degreeRouter.get('/get-all', degreeController.getAllCds); 
-
-
-module.exports = degreeRouter;
