+++ server/Services/Applications.js
@@ -1,16 +1,5 @@
 'use strict';
-
-/**
- * DAO for Application table
- * */
-
-const sqlite = require('sqlite3');
-
-// Open the database connection
-const db = new sqlite.Database('./Database/DB.sqlite', (err) => {
-    if (err) console.error(err.message);
-    console.log('Application DAO ready.');
-});
+import db from "../Database/DAO";
 
 module.exports = {
     retrieveKeywordsDetails: function (proposal_id) {
@@ -265,4 +254,62 @@ module.exports = {
             }
         })
     },
+
+    createApplication: function (proposalId, studentId) {
+        return new Promise((resolve, reject) => {
+            try {
+                console.log(proposalId, studentId);
+                db.run(`INSERT INTO Thesis_Applications (Th_Proposal_Id, Student_Id, Status)
+                                       VALUES (?, ?, "Pending")`, [proposalId, studentId], (err) => {
+                    if (err) reject(err);
+                    resolve();
+                });
+            } catch (error) {
+                reject(error);
+            }
+        });
+    },
+
+    acceptApplication: function (reqBody) {
+        return new Promise((resolve, reject) => {
+            try {
+                db.run(`UPDATE Thesis_Applications
+                        SET Status = "Accepted", Progress = "Active"
+                        WHERE Th_Proposal_Id = $proposalId AND Student_Id = $studentId;
+                        
+                        UPDATE Thesis_Applications
+                        SET Status = "Rejected"
+                        WHERE (Th_Proposal_Id = $proposalId AND Student_Id != $studentId) OR (Th_Proposal_Id != $proposalId AND Student_Id = $studentId)`,
+                    {
+                        $proposalId: reqBody.roposalId,
+                        $studentId: reqBody.studentId
+                    }, (err) => {
+                        if (err) reject(err);
+                        resolve();
+                    });
+            } catch (error) {
+                reject(error);
+            }
+        });
+    },
+
+
+    rejectApplication: function (reqBody) {
+        return new Promise((resolve, reject) => {
+            try {
+                db.run(`UPDATE Thesis_Applications
+                        SET Status = "Accepted", Progress = "Rejected"
+                        WHERE Th_Proposal_Id = $proposalId AND Student_Id = $studentId`,
+                    {
+                        $proposalId: reqBody.roposalId,
+                        $studentId: reqBody.studentId
+                    }, (err) => {
+                        if (err) reject(err);
+                        resolve();
+                    });
+            } catch (error) {
+                reject(error);
+            }
+        });
+    }
 }
