+++ server/fileList.txt
@@ -1,0 +1,11 @@
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/ProposalRouter.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterAPI.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterApplication.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterAuth.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterCoSupervisor.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterDegree.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterResearchGroup.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterSearch.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterTeacher.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterTest.js
+/Users/mahmoud/Repos/thesis-management-team17/server/Router/RouterThesis.js
